Move event callback handling to TransactionStatusHandler
